+++ nuxt.config.js
@@ -5,7 +5,7 @@ const { MICRO_CMS_KEY, ARTICLE_URL, TAG_URL, CATEGORY_URL } = process.env;
 
 export default {
   telemetry: false,
-  mode: 'spa',
+  mode: 'spa', // 'universal',
   generate: {
     routes() {
       // タグのルーティング
