+++ pages/category/_key/index.vue
@@ -11,52 +11,18 @@
 </template>
 
 <script>
-// FIX:後でresult->tagに変更(pagesのresult/_keyも修正)
-const POSTS_PER_PAGE = 10;
+import cms from 'assets/js/mixin/cms.mixin.js';
 
 export default {
   name: 'Category',
+  mixins: [cms],
   data() {
     return {
-      page: 0,
-      posts: [],
-      isLoad: true
+      fields: 'id,title,summary,tags,category,createdAt,updatedAt',
+      filter: `category[equals]${this.$route.query.id}`,
+      url: process.env.CATEGORY_URL,
+      posts: []
     };
-  },
-  computed: {
-    pageIndex() {
-      return this.page * POSTS_PER_PAGE;
-    }
-  },
-  methods: {
-    async infiniteHandler($state) {
-      if (this.isLoad) {
-        // 指定のタグIDを含む記事を探す
-        const FILTER_CATEGORY = `category[equals]${this.$route.query.id}`;
-        // クエリ
-        const OPTIONS = {
-          fields: 'id,title,summary,tags,category,createdAt,updatedAt',
-          filters: FILTER_CATEGORY,
-          limit: POSTS_PER_PAGE,
-          offset: this.pageIndex
-        };
-        // コンテンツの取得
-        const contents = await this.$getContents(
-          process.env.MICRO_CMS_KEY,
-          process.env.ARTICLE_URL,
-          OPTIONS
-        );
-        // ページング
-        if (contents.length > 0) {
-          this.page++;
-          this.posts.push(...contents);
-          $state.loaded();
-        } else {
-          $state.complete();
-          this.isLoad = false;
-        }
-      }
-    }
   }
 };
 </script>
