+++ plugins/components.js
@@ -1,12 +1,10 @@
 import Vue from 'vue';
 import InfiniteLoading from 'vue-infinite-loading';
-import Header from '@/components/Header';
 import NaviBar from '@/components/NaviBar';
 import AricleList from '@/components/ArticleList';
 import layout from './layout.css';
 
 Vue.use(layout);
 Vue.use(InfiniteLoading);
-Vue.component('Header', Header);
 Vue.component('NaviBar', NaviBar);
 Vue.component('ArticleList', AricleList);
